# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "SubscriptionDto": {
        "properties": {
          "currency": {
            "enum": [
              "USD",
              "KZT"
            ],
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "periodDays": {
            "format": "int32",
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "EXPIRED"
            ],
            "type": "string"
          },
          "subscriptionId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Keycloak-issued access token. Two realms are accepted, the mobile realm and the admin realm; the token's issuer selects which. Endpoints under /admin additionally require the ADMIN role.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Yogida Team",
      "url": "https://github.com/yogida"
    },
    "description": "REST API for the Yogida meditation platform. \nProvides presigned Cloudflare R2 streaming URLs and a catalog of available meditation media grouped by bucket.",
    "title": "Meditation Service API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/s3/buckets/{bucketName}": {
      "delete": {
        "description": "Bucket must be empty before deletion.",
        "operationId": "deleteBucket",
        "parameters": [
          {
            "description": "Bucket name",
            "in": "path",
            "name": "bucketName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bucket deleted"
          },
          "404": {
            "description": "Bucket not found"
          },
          "409": {
            "description": "Bucket is not empty"
          }
        },
        "summary": "Delete a bucket",
        "tags": [
          "Admin — S3 Buckets"
        ]
      },
      "post": {
        "operationId": "createBucket",
        "parameters": [
          {
            "description": "Bucket name",
            "in": "path",
            "name": "bucketName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Bucket created"
          },
          "400": {
            "description": "Invalid bucket name or bucket already exists"
          }
        },
        "summary": "Create a bucket",
        "tags": [
          "Admin — S3 Buckets"
        ]
      }
    },
...
    "/subscriptions": {
      "get": {
        "operationId": "listSubscriptionPlans",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Subscription plans retrieved successfully"
          }
        },
        "summary": "Get all subscription plans",
        "tags": [
          "Subscriptions"
        ]
      },
      "post": {
        "operationId": "createSubscriptionPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan created successfully"
          }
        },
        "summary": "Create a new subscription plan",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/subscriptions/search": {
      "get": {
        "operationId": "findSubscriptionPlanByName",
        "parameters": [
          {
            "description": "Subscription name",
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan not found"
          }
        },
        "summary": "Find subscription plan by name",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/subscriptions/{id}": {
      "delete": {
        "operationId": "deleteSubscriptionPlan",
        "parameters": [
          {
            "description": "Subscription ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subscription plan deleted successfully"
          },
          "404": {
            "description": "Subscription plan not found"
          }
        },
        "summary": "Delete a subscription plan",
        "tags": [
          "Subscriptions"
        ]
      },
      "get": {
        "operationId": "getSubscriptionPlanById",
        "parameters": [
          {
            "description": "Subscription ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan not found"
          }
        },
        "summary": "Get subscription plan by ID",
        "tags": [
          "Subscriptions"
        ]
      },
      "put": {
        "operationId": "updateSubscriptionPlan",
        "parameters": [
          {
            "description": "Subscription ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan updated successfully"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            },
            "description": "Subscription plan not found"
          }
        },
        "summary": "Update an existing subscription plan",
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}